+++ booksim2/routers/vct_router.cpp
@@ -424,10 +424,10 @@ namespace Booksim
       _in_queue_flits.clear();
 
       while(!_proc_credits.empty()) {
-
-        pair<long, pair<Credit *, int> > const & item = _proc_credits.front();
-
-        long const time = item.first;
+        //BSMOD: Change time to long long
+        pair<long long, pair<Credit *, int> > const & item = _proc_credits.front();
+        //BSMOD: Change time to long long
+        long long const time = item.first;
         if(GetSimTime() < time) {
           break;
         }
@@ -465,12 +465,12 @@ namespace Booksim
     void VCTRouter::_RouteEvaluate( )
     {
       assert(_routing_delay);
-
-      for(deque<pair<long, pair<int, int> > >::iterator iter = _route_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<int, int> > >::iterator iter = _route_vcs.begin();
           iter != _route_vcs.end();
           ++iter) {
-        
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         if(time >= 0) {
           break;
         }
@@ -505,10 +505,10 @@ namespace Booksim
       assert(_routing_delay);
 
       while(!_route_vcs.empty()) {
-
-        pair<long, pair<int, int> > const & item = _route_vcs.front();
-
-        long const time = item.first;
+        //BSMOD: Change time to long long
+        pair<long long, pair<int, int> > const & item = _route_vcs.front();
+        //BSMOD: Change time to long long
+        long long const time = item.first;
         if((time < 0) || (GetSimTime() < time)) {
           break;
         }
@@ -559,12 +559,12 @@ namespace Booksim
       assert(_vc_allocator);
 
       bool watched = false;
-
-      for(deque<pair<long, pair<pair<int, int>, int> > >::iterator iter = _vc_alloc_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<pair<int, int>, int> > >::iterator iter = _vc_alloc_vcs.begin();
           iter != _vc_alloc_vcs.end();
           ++iter) {
-
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         if(time >= 0) {
           break;
         }
@@ -716,11 +716,12 @@ namespace Booksim
         _vc_allocator->PrintGrants( gWatchOut );
       }
 
-      for(deque<pair<long, pair<pair<int, int>, int> > >::iterator iter = _vc_alloc_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<pair<int, int>, int> > >::iterator iter = _vc_alloc_vcs.begin();
           iter != _vc_alloc_vcs.end();
           ++iter) {
-
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         if(time >= 0) {
           break;
         }
@@ -786,11 +787,12 @@ namespace Booksim
         return;
       }
 
-      for(deque<pair<long, pair<pair<int, int>, int> > >::iterator iter = _vc_alloc_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<pair<int, int>, int> > >::iterator iter = _vc_alloc_vcs.begin();
           iter != _vc_alloc_vcs.end();
           ++iter) {
-        
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         assert(time >= 0);
         if(GetSimTime() < time) {
           break;
@@ -853,10 +855,10 @@ namespace Booksim
       assert(_vc_allocator);
 
       while(!_vc_alloc_vcs.empty()) {
-
-        pair<long, pair<pair<int, int>, int> > const & item = _vc_alloc_vcs.front();
-
-        long const time = item.first;
+        //BSMOD: Change time to long long
+        pair<long long, pair<pair<int, int>, int> > const & item = _vc_alloc_vcs.front();
+        //BSMOD: Change time to long long
+        long long const time = item.first;
         if((time < 0) || (GetSimTime() < time)) {
           break;
         }
@@ -943,11 +945,12 @@ namespace Booksim
     {
       assert(_hold_switch_for_packet);
 
-      for(deque<pair<long, pair<pair<int, int>, int> > >::iterator iter = _sw_hold_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<pair<int, int>, int> > >::iterator iter = _sw_hold_vcs.begin();
           iter != _sw_hold_vcs.end();
           ++iter) {
-        
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         if(time >= 0) {
           break;
         }
@@ -1018,10 +1021,10 @@ namespace Booksim
       assert(_hold_switch_for_packet);
 
       while(!_sw_hold_vcs.empty()) {
-        
-        pair<long, pair<pair<int, int>, int> > const & item = _sw_hold_vcs.front();
-        
-        long const time = item.first;
+        //BSMOD: Change time to long long
+        pair<long long, pair<pair<int, int>, int> > const & item = _sw_hold_vcs.front();
+        //BSMOD: Change time to long long
+        long long const time = item.first;
         if(time < 0) {
           break;
         }
@@ -1338,12 +1341,12 @@ namespace Booksim
     void VCTRouter::_SWAllocEvaluate( )
     {
       bool watched = false;
-
-      for(deque<pair<long, pair<pair<int, int>, int> > >::iterator iter = _sw_alloc_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<pair<int, int>, int> > >::iterator iter = _sw_alloc_vcs.begin();
           iter != _sw_alloc_vcs.end();
           ++iter) {
-
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         if(time >= 0) {
           break;
         }
@@ -1508,12 +1511,12 @@ namespace Booksim
           _spec_sw_allocator->PrintGrants(gWatchOut);
         }
       }
-      
-      for(deque<pair<long, pair<pair<int, int>, int> > >::iterator iter = _sw_alloc_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<pair<int, int>, int> > >::iterator iter = _sw_alloc_vcs.begin();
           iter != _sw_alloc_vcs.end();
           ++iter) {
-
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         if(time >= 0) {
           break;
         }
@@ -1649,12 +1652,12 @@ namespace Booksim
       if(!_speculative && (_sw_alloc_delay <= 1)) {
         return;
       }
-
-      for(deque<pair<long, pair<pair<int, int>, int> > >::iterator iter = _sw_alloc_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<pair<int, int>, int> > >::iterator iter = _sw_alloc_vcs.begin();
           iter != _sw_alloc_vcs.end();
           ++iter) {
-
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         assert(time >= 0);
         if(GetSimTime() < time) {
           break;
@@ -1853,10 +1856,10 @@ namespace Booksim
     void VCTRouter::_SWAllocUpdate( )
     {
       while(!_sw_alloc_vcs.empty()) {
-
-        pair<long, pair<pair<int, int>, int> > const & item = _sw_alloc_vcs.front();
-
-        long const time = item.first;
+        //BSMOD: Change time to long long
+        pair<long long, pair<pair<int, int>, int> > const & item = _sw_alloc_vcs.front();
+        //BSMOD: Change time to long long
+        long long const time = item.first;
         if((time < 0) || (GetSimTime() < time)) {
           break;
         }
@@ -2151,11 +2154,12 @@ namespace Booksim
 
     void VCTRouter::_SwitchEvaluate( )
     {
-      for(deque<pair<long, pair<Flit *, pair<int, int> > > >::iterator iter = _crossbar_flits.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<Flit *, pair<int, int> > > >::iterator iter = _crossbar_flits.begin();
           iter != _crossbar_flits.end();
           ++iter) {
-        
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         if(time >= 0) {
           break;
         }
